updated file

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
 */
 int main(void)
 {
-	printf("Size of a char: %u byete(s)\n", (unsigned long)sizeof(char));
-	printf("Size of an int: %u byete(s)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %u byete(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %u byete(s)\n", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %u byete(s)\n", (unsigned long)sizeof(float));
+	printf("Size of a char: %u byete(s)\n", sizeof(char));
+	printf("Size of an int: %u byete(s)\n", sizeof(int));
+	printf("Size of a long int: %u byete(s)\n", sizeof(long int));
+	printf("Size of a long long int: %u byete(s)\n", sizeof(long long int));
+	printf("Size of a float: %u byete(s)\n", sizeof(float));
 	return (0);
 }
